pyproject.toml
[project]
name = "marunochithe"
version = "0.1.0"
description = "The most powerful self-hosted coding assistant"
authors = [{name = "Cesar", email = "cesar@marunochithe.ai"}]
requires-python = ">=3.9"
dependencies = [
    # Core framework
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",

    # HTTP clients
    "httpx>=0.26.0",
    "aiohttp>=3.9.0",

    # Ollama integration
    "ollama>=0.1.6",
    "litellm>=1.20.0",

    # Database & RAG
    "chromadb>=0.4.22",
    "aiosqlite>=0.19.0",

    # Code understanding
    "tree-sitter>=0.20.4",
    "tree-sitter-python>=0.20.4",
    "tree-sitter-javascript>=0.20.3",
    "tree-sitter-typescript>=0.20.5",

    # File watching
    "watchdog>=4.0.0",

    # CLI & TUI
    "typer[all]>=0.9.0",
    "rich>=13.7.0",
    "textual>=0.48.0",

    # Git integration
    "pygit2>=1.13.0",
    "gitpython>=3.1.40",

    # Fine-tuning (optional, for Phase 6)
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "bitsandbytes>=0.41.0",
    "datasets>=2.16.0",
    "trl>=0.7.0",

    # Utils
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

lsp = [
    "pygls>=1.3.0",
]

[project.scripts]
marunochithe = "marunochithe.integrations.cli:app"
marunochithe-server = "marunochithe.api.server:start"
marunochithe-tui = "marunochithe.integrations.tui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
